rename regibox to regybox - rename everything
