+++ .github/workflows/scheduled_holiday_runs.yml
@@ -1,5 +1,5 @@
 ---
-name: Trigger Regibox Holidays
+name: Trigger Regybox Holidays
 on:  # yamllint disable-line rule:truthy
   workflow_dispatch:
   schedule:  # 48 hours in advance
@@ -30,7 +30,7 @@ jobs:
         run: |
           echo "WEEKDAY=$([[ "$(date -d '+2 days' +%u)" -lt 6 ]] && echo "true" || echo "false")" >> "$GITHUB_OUTPUT"
   build:
-    name: Trigger Regibox run
+    name: Trigger Regybox run
     runs-on: ubuntu-latest
     needs:
       - weekday
@@ -54,8 +54,8 @@ jobs:
           echo "CLASS_DATE=$(date -d '+2 days' +%F)" >> $GITHUB_ENV
           echo "CLASS_TIME='10:00'" >> $GITHUB_ENV
           echo "CLASS_TYPE='Weekend WOD Rato'" >> $GITHUB_ENV
-      - name: Regibox enroll
-        run: pants run regibox:main -- ${{ env.CLASS_DATE }} ${{ env.CLASS_TIME }} ${{ env.CLASS_TYPE }}
+      - name: Regybox enroll
+        run: pants run regybox:main -- ${{ env.CLASS_DATE }} ${{ env.CLASS_TIME }} ${{ env.CLASS_TYPE }}
         env:
           PHPSESSID: ${{ secrets.PHPSESSID }}
           REGIBOX_USER: ${{ secrets.REGIBOX_USER }}
@@ -68,7 +68,7 @@ jobs:
           username: ${{ secrets.EMAIL_USERNAME }}
           password: ${{ secrets.EMAIL_PASSWORD }}
           subject: >
-            Regibox HOLIDAY Auto-enroll: ${{ job.status }} for ${{ env.CLASS_TYPE }} on ${{ env.CLASS_DATE }}
+            Regybox HOLIDAY Auto-enroll: ${{ job.status }} for ${{ env.CLASS_TYPE }} on ${{ env.CLASS_DATE }}
             at ${{ env.CLASS_TIME }}
           body: >
             Job '${{ github.job }}' in worflow ${{ github.workflow }} of ${{ github.repository }} finished in a
@@ -76,4 +76,4 @@ jobs:
 
             See ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}.
           to: m@rtimlobao.com
-          from: Regibox Auto-enroll
+          from: Regybox Auto-enroll
