+++ README.md
@@ -1,1 +1,1 @@
-# regibox
+# regybox
