+++ regybox/__main__.py
@@ -1,7 +1,7 @@
 import sys
 
-from regibox.common import LOGGER
-from regibox.regibox import main
+from regybox.common import LOGGER
+from regybox.regybox import main
 
 if __name__ == "__main__":
     class_date, class_time, class_type = sys.argv[1:]  # pylint: disable=unbalanced-tuple-unpacking
