+++ regybox/utils/singleton.py
